Break higher-order choreography tests out into own file
